dsa-cli/requirements.txt
boto3
python-dotenv
typer